separated the Card Component from the List
