+++ src/components/List/styles.ts
@@ -9,31 +9,4 @@ export const useStyles = createStyles((theme) => ({
     margin: `${theme.spacing.xl}px auto`,
     gridGap: 20,
   },
-  card: {
-    borderWidth: 2,
-    borderColor: theme.primaryColor,
-    maxWidth: theme.breakpoints.sm,
-    width: "90%",
-    minWidth: 385,
-  },
-  longURL: {
-    color:
-      theme.colorScheme === "dark"
-        ? theme.colors.gray[5]
-        : theme.colors.gray[6],
-  },
-  shortURL: {
-    textDecoration: "underline",
-    textUnderlineOffset: 3,
-  },
-  shortURLContainer: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  redirectsStats: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-  },
 }));
